refactoring the gui node builder
